split out listen mod
